backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-generativeai
requests
